requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
lxml==4.3.0
MarkupSafe==1.1.0
pdfdocument==3.2
pdfminer.six==20181108
pycryptodome==3.7.2
PyPDF2==1.26.0
python-docx==0.8.10
six==1.12.0
sortedcontainers==2.1.0
Werkzeug==0.14.1
